# openapi.yaml
openapi: 3.1.0
info:
  title: examOnline API
  description: 在线考试系统后端 API 契约。用于前端 @hey-api/openapi-ts 生成类型化 axios 客户端。契约如实描述当前 14
    个 Controller 的现状，不改任何签名/DTO/逻辑。新增端点后必须重新运行导出命令更新 openapi.yaml。
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- Authorization: []
paths:
...
  /api/papers/{id}:
    get:
      tags:
      - paper-controller
      operationId: detail_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperDetailResponse"
    put:
      tags:
      - paper-controller
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperDetailResponse"
    delete:
      tags:
      - paper-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/papers/{id}/questions/{questionId}/score:
    put:
      tags:
      - paper-controller
      operationId: updateQuestionScore
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: questionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaperQuestionScoreRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/papers/{id}/questions/order:
    put:
      tags:
      - paper-controller
      operationId: updateOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperDetailResponse"
...
  /api/papers:
    get:
      tags:
      - paper-controller
      operationId: page_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListPaperResponse"
    post:
      tags:
      - paper-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperResponse"
  /api/papers/{id}/snapshot:
    get:
      tags:
      - paper-controller
      operationId: getSnapshot
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperSnapshotResponse"
    post:
      tags:
      - paper-controller
      operationId: generateSnapshot
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperSnapshotResponse"
  /api/papers/{id}/questions:
    post:
      tags:
      - paper-controller
      operationId: addQuestion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPaperQuestionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperQuestionItemResponse"
  /api/papers/{id}/questions/random:
    post:
      tags:
      - paper-controller
      operationId: commitRandomDraw
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RandomDrawRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperDetailResponse"
  /api/papers/{id}/questions/batch:
    post:
      tags:
      - paper-controller
      operationId: addQuestions
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchAddPaperQuestionsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaperDetailResponse"
  /api/papers/random-draw/preview:
    post:
      tags:
      - paper-controller
      operationId: previewDraw
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RandomDrawRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRandomDrawPreviewResponse"
...
  /api/exams/{id}/snapshot:
    get:
      tags:
      - exam-controller
      operationId: getSnapshot_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseExamSnapshotResponse"
...
  /api/papers/{id}/questions/{questionId}:
    delete:
      tags:
      - paper-controller
      operationId: removeQuestion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: questionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
components:
  schemas:
...
    PaperUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 128
          minLength: 0
        description:
          type: string
          maxLength: 512
          minLength: 0
        totalScore:
          type: number
          maximum: 9999.9
          minimum: 0
    ApiResponsePaperDetailResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PaperDetailResponse"
    PaperDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        totalScore:
          type: number
        questionCount:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        snapshotId:
          type: integer
          format: int64
        createdBy:
          type: integer
          format: int64
        createdTime:
          type: string
          format: date-time
        updatedTime:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/PaperQuestionItemResponse"
    PaperQuestionItemResponse:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        score:
          type: number
        questionType:
          type: integer
          format: int32
        content:
          type: string
        choices:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        defaultScore:
          type: number
        questionDeleted:
          type: boolean
    UpdatePaperQuestionScoreRequest:
      type: object
      properties:
        score:
          type: number
          maximum: 999.9
          minimum: 0.5
      required:
      - score
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data: {}
    PaperOrderRequest:
      type: object
      properties:
        questionIds:
          type: array
          items:
            type: integer
            format: int64
          minItems: 1
      required:
      - questionIds
...
    JsonNode: {}
...
    PaperCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 128
          minLength: 0
        description:
          type: string
          maxLength: 512
          minLength: 0
        totalScore:
          type: number
          maximum: 9999.9
          minimum: 0
      required:
      - title
      - totalScore
    ApiResponsePaperResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PaperResponse"
    PaperResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        totalScore:
          type: number
        questionCount:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        snapshotId:
          type: integer
          format: int64
        createdBy:
          type: integer
          format: int64
        createdTime:
          type: string
          format: date-time
    ApiResponsePaperSnapshotResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PaperSnapshotResponse"
    PaperSnapshotResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        paperId:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        totalScore:
          type: number
        questionCount:
          type: integer
          format: int32
        content:
          $ref: "#/components/schemas/JsonNode"
        createdTime:
          type: string
          format: date-time
    AddPaperQuestionRequest:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
        score:
          type: number
          maximum: 999.9
          minimum: 0.5
      required:
      - questionId
    ApiResponsePaperQuestionItemResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/PaperQuestionItemResponse"
    RandomDrawRequest:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
          minItems: 1
      required:
      - rules
    Rule:
      type: object
      properties:
        type:
          type: integer
          format: int32
          maximum: 4
          minimum: 1
        difficulty:
          type: integer
          format: int32
          maximum: 3
          minimum: 1
        tagIds:
          type: array
          items:
            type: integer
            format: int64
        count:
          type: integer
          format: int32
          minimum: 1
      required:
      - count
    BatchAddPaperQuestionsRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
          maxItems: 100
          minItems: 0
      required:
      - items
    Item:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
        score:
          type: number
          maximum: 999.9
          minimum: 0.5
      required:
      - questionId
    ApiResponseRandomDrawPreviewResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/RandomDrawPreviewResponse"
    DrawnQuestion:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: integer
          format: int32
        typeName:
          type: string
        content:
          type: string
        score:
          type: number
        difficulty:
          type: integer
          format: int32
    RandomDrawPreviewResponse:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/RuleResult"
        total:
          type: integer
          format: int32
    RuleResult:
      type: object
      properties:
        ruleIndex:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
        questions:
          type: array
          items:
            $ref: "#/components/schemas/DrawnQuestion"
...
    ApiResponseListPaperResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PaperResponse"
...
    ApiResponseExamSnapshotResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/ExamSnapshotResponse"
    ExamSnapshotResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        examId:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
        exam:
          $ref: "#/components/schemas/JsonNode"
        paper:
          $ref: "#/components/schemas/JsonNode"
        createdTime:
          type: string
          format: date-time
...
  securitySchemes:
    Authorization:
      type: http
      description: JWT Access Token，格式：Bearer <token>
      name: Authorization
      scheme: bearer
      bearerFormat: JWT